created PSO but not tested

+++ ctypes1/clibrary.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 // Function that will take in a string and int and display them in message and return a string
-char* display(char* str, int age){
+char* display(char* str, double age){
     printf("my name is %s, and my age is %d\n", str, age);
     return "completed";
 }
